Added admin functionality for Animal; tests update

+++ PetKeeperBackend/Services/AnimalManagementService.cs
@@ -1,15 +1,8 @@
 ﻿using Grpc.Core;
-using grpc_hello_world;
 using grpc_hello_world.Models;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.AspNetCore.Identity;
-using System.Drawing.Printing;
-using System.Reflection;
-using System.Security.Policy;
 using Microsoft.AspNetCore.Authorization;
 using System.Security.Claims;
-using System.Reflection.Metadata.Ecma335;
-using Azure.Core;
 
 namespace grpc_hello_world.Services
 {
@@ -29,14 +22,18 @@ namespace grpc_hello_world.Services
         {
             var userContext = context.GetHttpContext().User;
             var ownerId = userContext.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var userRole = userContext.FindFirst(ClaimTypes.Role)?.Value;
+
+            if (request.HasOwnerId && userRole == "Admin")
+                ownerId = request.OwnerId;
 
             var animal = new Animal
             {
                 Name = request.Name,
                 OwnerId = Guid.Parse(ownerId),
                 Type = request.Type,
-                Description = request.Description ?? "",
-                Photos = request.Photos.ToArray()
+                Description = request.Description,
+                Photo = request.Photo
             };
 
             try
@@ -55,6 +52,9 @@ namespace grpc_hello_world.Services
         [Authorize]
         public override async Task<AnimalList> GetAnimals(AnimalGet request, ServerCallContext context)
         {
+            var userContext = context.GetHttpContext().User;
+            var userId = userContext.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
             var query = _context.Animals.AsQueryable();
 
             if (request.HasId)
@@ -73,10 +73,16 @@ namespace grpc_hello_world.Services
                 query = query.Where(a => words.Any(word => a.Description.Contains(word)));
             }
 
+            if (request.HasOwnerId)
+                if (string.IsNullOrEmpty(request.OwnerId))
+                    query = query.Where(a => a.OwnerId.ToString() == userId);
+                else
+                    query = query.Where(a => a.OwnerId.ToString() == request.OwnerId);
+
             var animals = await query.ToListAsync();
 
-            var addressList = new AnimalList();
-            addressList.Animals.AddRange(animals.Select(a => new AnimalCreate
+            var animalList = new AnimalList();
+            animalList.Animals.AddRange(animals.Select(a => new AnimalCreate
             {
                 Id = a.Id.ToString(),
                 OwnerId = a.OwnerId.ToString(),
@@ -85,7 +91,7 @@ namespace grpc_hello_world.Services
                 Description = a.Description ?? ""
             }));
 
-            return addressList;
+            return animalList;
 
         }
 
@@ -94,17 +100,29 @@ namespace grpc_hello_world.Services
         {
             var userContext = context.GetHttpContext().User;
             var ownerId = userContext.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var userRole = userContext.FindFirst(ClaimTypes.Role)?.Value;
 
             try
             {
-                Animal? animal = await _context.Animals.SingleOrDefaultAsync(
-                    a => a.Id.ToString() == request.Id && a.OwnerId.ToString() == ownerId)
-                    ?? throw new RpcException(new Status(StatusCode.NotFound, "Animal does not exist"));
+                Animal? animal;
+                if (userRole == "Admin")
+                {
+                    animal = await _context.Animals.SingleOrDefaultAsync(
+                    a => a.Id.ToString() == request.Id);
+                }
+                else
+                {
+                    animal = await _context.Animals.SingleOrDefaultAsync(
+                    a => a.Id.ToString() == request.Id && a.OwnerId.ToString() == ownerId);
+                }
+
+                if (animal == null)
+                    throw new RpcException(new Status(StatusCode.NotFound, "Animal does not exist"));
 
                 _context.Animals.Remove(animal);
                 await _context.SaveChangesAsync();
 
-                return new AnimalMinimal { Id = animal.Id.ToString(), OwnerId = animal.OwnerId.ToString()  };
+                return new AnimalMinimal { Id = animal.Id.ToString(), OwnerId = animal.OwnerId.ToString() };
             }
             catch (DbUpdateException e)
             {
@@ -119,10 +137,25 @@ namespace grpc_hello_world.Services
 
             var userContext = context.GetHttpContext().User;
             var ownerId = userContext.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-            Animal? animal = await _context.Animals.FirstOrDefaultAsync(
-                a => a.Id.ToString() == request.Id && a.OwnerId.ToString() == ownerId)
-                ??
+            var userRole = userContext.FindFirst(ClaimTypes.Role)?.Value;
+
+            Animal? animal;
+            if (userRole == "Admin")
+            {
+                animal = await _context.Animals.FirstOrDefaultAsync(
+                    a => a.Id.ToString() == request.Id);
+                
+            }
+            else
+            {
+                animal = await _context.Animals.FirstOrDefaultAsync(
+                    a => a.Id.ToString() == request.Id && a.OwnerId.ToString() == ownerId);
+            }
+            if (animal == null)
+            {
                 throw new RpcException(new Status(StatusCode.NotFound, "Animal not found!"));
+            }
+            
 
             var requestProperties = request_type.GetProperties();
             var modifiedProperties = new HashSet<string>();
